Remove unused code part 1 (#95)
* remove temporary balancer alert

* remove exploit watcher

* DefaultLayout: remove unused code & move it

* remove unused hero(e)s components

* remove unused globalstats component

* images: remove unused code & move it

* navs: remove unused code & move it WIP

* fix relative path

* fix extensions

* nothing to setup or return

* remove more unused code WIP

* restore original file for now
