--- src/beethovenx/components/images/AppIcon.vue
@@ -1,19 +1,0 @@
-<template>
-  <img
-    src="~@/beethovenx/assets/images/logo.svg"
-    width="175"
-    class="pl-2 xl:pl-6 "
-  />
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-  name: 'AppIcon',
-
-  setup() {
-    return {};
-  }
-});
-</script>
