+++ src/locales/default.json
@@ -301,7 +301,7 @@
   "lpsEarnFee": "<span>LPs earn <b>{0}</b> in fees</span>",
   "managedPool": "Managed pool",
   "manageLists": "Manage lists",
-  "marketConditionsWarning": "Market conditions may change between the time your order is submitted and the time it gets executed on Ethereum. Slippage tolerance is the maximum change in price you are willing to accept. This protects you from front-running bots and miner extractable value (MEV).",
+  "marketConditionsWarning": "Market conditions may change between the time your order is submitted and the time it gets executed on Fantom. Slippage tolerance is the maximum change in price you are willing to accept. This protects you from front-running bots and miner extractable value (MEV).",
   "max": "Max",
   "maxed": "Maxed",
   "maximum": "Maximum",
